Column position directly calculated
I FINALLY found a decent equation for ray distance -> column height, but in the process I gained the ability to just calculate the column height directly.

Now you can set the height for the player as well as the walls. You can also have the player look up and down (which at this point effectively just scrolls the screen up and down...)

At the moment there is plenty of cleaning up to do before I move on to other features